pyproject.toml
[tool.poetry]
name = "tour-site"
version = "0.1.0"
description = ""
authors = ["Nurlan-Aliev <nualiev612@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.5"}
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
redis = "^5.2.0"
pytest = "^8.3.4"
pytest-cov = "^6.0.0"


[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
